Move automateWorkBoats to WorkerAutomation - for readability and better chances to create synergies
